Removed download of pre-trained models from automatically run tests (#200)
This PR removes the training with pre-trained models from the RL tests. 
However, a further GitHub workflow is added to manually trigger those
tests. That possibility to trigger that file is apparently only shown
after merging this PR into `main`.

While the test time is only reduced by roughly 20%, it allows to not
have failed tests when changing anything related to the RL model (such
as a different number of compilation passes).